src/fiqat/methods/fiqa/faceqnet/requirements.txt
numpy<1.20.0,>=1.18.5
tensorflow-cpu>=2.5.3
